feat: remove useless code for tokenizer_tools
