--- seq2annotation/helper/collect_label_to_file.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-from typing import List
-
-from tokenizer_tools.conllz.tag_collector import collect_label_to_file
-
-
-def collect_label_to_file(input_file_list: List[str], output_file: str):
-    collect_label_to_file(input_file_list, output_file)
