--- seq2annotation/trainer/cli_paddle.py
@@ -1,4 +1,0 @@
-from seq2annotation.trainer.paddle_train import Train
-
-train = Train()
-train.train()
